--- src/store/actions.js
@@ -1,5 +1,0 @@
-import { createAction } from "@reduxjs/toolkit"
-
-// Actions that are common to multiple slices.
-
-export const urlChanged = createAction("urlChanged")
