+++ src/store/actions.ts
@@ -1,0 +1,9 @@
+import { createAction } from "@reduxjs/toolkit"
+
+// Actions that are common to multiple slices.
+interface UrlChangedPayload {
+  search?: string
+  filters?: Record<string, string[]>
+}
+
+export const urlChanged = createAction<UrlChangedPayload>("urlChanged")
